# docs/swagger/swagger.yaml
definitions:
  handlers.addAgendaItemRequest:
    properties:
      description:
        type: string
      duration_minutes:
        example: 15
        type: integer
      title:
        example: Цели спринта
        type: string
    type: object
...
  handlers.agendaItemResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      event_id:
        type: integer
      id:
        type: integer
      is_done:
        type: boolean
      position:
        type: integer
      title:
        type: string
    type: object
  handlers.createEventRequest:
    properties:
      description:
        example: Расставляем приоритеты
        type: string
      end_time:
        example: "2026-05-01T11:00:00Z"
        type: string
      location:
        example: Переговорная №2
        type: string
      meeting_url:
        example: https://meet.google.com/abc-defg-hij
        type: string
      start_time:
        example: "2026-05-01T10:00:00Z"
        type: string
      status:
        enum:
        - scheduled
        - cancelled
        - completed
        example: scheduled
        type: string
      title:
        example: Планирование спринта
        type: string
    type: object
...
  handlers.eventResponse:
    properties:
      created_at:
        type: string
      creator_id:
        type: integer
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      location:
        type: string
      meeting_url:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  handlers.updateAgendaItemRequest:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      is_done:
        type: boolean
      title:
        type: string
    type: object
  handlers.updateEventRequest:
    properties:
      description:
        type: string
      end_time:
        type: string
      location:
        type: string
      meeting_url:
        type: string
      start_time:
        type: string
      status:
        enum:
        - scheduled
        - cancelled
        - completed
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Сервис для управления рабочими мероприятиями.
  title: Meeting Service API
  version: "1.0"
paths:
  /events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.eventResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Список всех встреч
      tags:
      - events
    post:
      consumes:
      - application/json
      parameters:
      - description: Параметры встречи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.createEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.eventResponse'
        "400":
          description: validation error
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать встречу
      tags:
      - events
  /events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: no content
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: event not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить встречу
      tags:
      - events
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.eventResponse'
        "404":
          description: event not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить встречу по id
      tags:
      - events
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Новые параметры
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.updateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.eventResponse'
        "400":
          description: validation error
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: event not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить встречу
      tags:
      - events
  /events/{id}/agenda:
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.agendaItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Список пунктов повестки
      tags:
      - agenda
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Пункт повестки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.addAgendaItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.agendaItemResponse'
        "400":
          description: validation error
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Добавить пункт повестки
      tags:
      - agenda
  /events/{id}/agenda/{agendaId}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Agenda Item ID
        in: path
        name: agendaId
        required: true
        type: integer
      responses:
        "204":
          description: no content
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: agenda item not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить пункт повестки
      tags:
      - agenda
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Agenda Item ID
        in: path
        name: agendaId
        required: true
        type: integer
      - description: Новые поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.updateAgendaItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.agendaItemResponse'
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: agenda item not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить пункт повестки
      tags:
      - agenda
...
swagger: "2.0"